refactor(renderer): keep LaunchTeamDialog under its frozen size cap
Move the one-time localStorage migration into its own module so the
dialog stays at its 3333-line frozen cap after the dirty-tracking fix.
